+++ src/entities/__init__.py
@@ -1,0 +1,9 @@
+from .sector import Sector
+from .player import PlayerCharacter
+from .inventory import PlayerInventory
+from .item import Item
+from .enemy import Enemy
+
+sector = Sector()
+player_character = PlayerCharacter()
+player_inventory = PlayerInventory()
