+++ src/entities/item.py
@@ -1,10 +1,10 @@
 from dataclasses import dataclass
+from random import choice
 from random import choices, choice, randint
+
+from display import char_print
+
 from constants import (
-    ENEMY_LISTS,
-    SECTOR_LIST,
-    SECTOR_WEIGHTS,
-    ENEMY_NAMES,
     ADJECTIVES,
     ITEM_NAMES,
     ITEM_TYPES_LIST,
@@ -12,122 +12,6 @@ from constants import (
     ATTIRE_INSPECT,
     CONSUMABLE_INSPECT,
 )
-from time import sleep
-
-
-def char_print(string):
-    for character in string[:-1]:
-        print(character, end="", flush=True)
-        sleep(0.02)
-    print(string[-1], flush=True)
-    sleep(1)
-
-
-def char_input(string, title=False, lower=False):
-    for character in string:
-        print(character, end="", flush=True)
-        sleep(0.02)
-    sleep(1)
-
-    if title:
-        return input().title()
-    elif lower:
-        return input().lower()
-    else:
-        return input()
-
-
-@dataclass
-class Sector:
-    current_sector: str = ""
-    current_weights: tuple = (1, 1, 1)
-
-    @staticmethod
-    def _get_weights_dict():
-        return {
-            "caves": SECTOR_WEIGHTS.CAVES,
-            "tombs": SECTOR_WEIGHTS.TOMBS,
-            "sewers": SECTOR_WEIGHTS.SEWERS,
-            "dungeons": SECTOR_WEIGHTS.DUNGEONS,
-        }
-
-    def roll_sector(self):
-        self.current_sector = choice(SECTOR_LIST)
-        self.current_weights = self._get_weights_dict()[self.current_sector]
-
-
-sector = Sector()
-
-
-@dataclass
-class Enemy:
-    category: str = ""
-    size: str = ""
-    rarity: str = ""
-    name: str = ""
-    health: int = 0
-    atk: int = 0
-    dmg: int = 0
-
-    def _assign_enemy_type(self):
-        self.category = "".join(
-            choices(ENEMY_LISTS.CATEGORY_LIST, weights=Sector.current_weights)
-        )
-        self.size = "".join(choices(ENEMY_LISTS.SIZE_LIST, weights=(3, 6, 1)))
-        self.rarity = "".join(choices(ENEMY_LISTS.RARITY_LIST, weights=(16, 3, 1)))
-
-    @staticmethod
-    def _get_name_group_dict():
-        return {
-            "small animal": ENEMY_NAMES.SMALL_ANIMAL,
-            "medium animal": ENEMY_NAMES.MEDIUM_ANIMAL,
-            "large animal": ENEMY_NAMES.LARGE_ANIMAL,
-            "small humanoid": ENEMY_NAMES.SMALL_HUMANOID,
-            "medium humanoid": ENEMY_NAMES.MEDIUM_HUMANOID,
-            "large humanoid": ENEMY_NAMES.LARGE_HUMANOID,
-            "small undead": ENEMY_NAMES.SMALL_UNDEAD,
-            "medium undead": ENEMY_NAMES.MEDIUM_UNDEAD,
-            "large undead": ENEMY_NAMES.LARGE_UNDEAD,
-        }
-
-    def _assign_name_to_enemy(self):
-        name_group = f"{self.size} {self.category}"
-        self.name = choice(self._get_name_group_dict()[name_group])
-
-    def _get_rarity_adjectives(self):
-        if self.rarity == "rare":
-            adjective: str = choice(ADJECTIVES.ENEMY_IS_RARE_ADJECTIVES[self.category])
-            self.name = f"{adjective} {self.name}"
-        if self.rarity == "boss":
-            adjective: str = choice(ADJECTIVES.ENEMY_IS_BOSS_ADJECTIVES[self.category])
-            self.name = f"{adjective} {self.name}"
-
-    def get_name(self):
-        return self.name
-
-    def _generate_enemy_stats(self):
-        stat_multiplier = 1
-        if self.size == "medium":
-            stat_multiplier = 2
-        elif self.size == "large":
-            stat_multiplier = 6
-        if self.rarity == "rare":
-            stat_multiplier += 2
-        elif self.rarity == "boss":
-            stat_multiplier += 6
-        self.health = randint(1, 6) * stat_multiplier
-        self.atk = int((randint(1, 6) * stat_multiplier) / 3)
-        self.dmg = int((randint(1, 6) * stat_multiplier) / 3)
-
-    def get_can_persuade(self):
-        if self.rarity != "boss" and self.category == "humanoid":
-            return True
-
-    def generate_enemy(self):
-        self._assign_enemy_type()
-        self._assign_name_to_enemy()
-        self._get_rarity_adjectives()
-        self._generate_enemy_stats()
 
 
 @dataclass
@@ -342,83 +226,3 @@ class Item:
             },
         }
         return dict_container
-
-
-@dataclass
-class PlayerCharacter:
-    name: str = ""
-    max_health: int = 0
-    health: int = 0
-    atk: int = 0
-    dmg: int = 0
-    dfc: int = 0
-    psd: int = 0
-
-    def generate_player_character_name(self):
-        self.name = char_input("What is your name?: ", title=True)
-
-    def generate_player_character_stats(self):
-        self.health += randint(7, 12)
-        self.max_health = self.health
-        self.atk += randint(0, 2)
-        self.dmg += randint(0, 2)
-        self.dfc += randint(0, 2)
-        self.psd += randint(0, 2)
-
-    def cap_health_at_max(self):
-        if self.health > self.max_health:
-            self.health = self.max_health
-
-    def print_player_character_stats(self):
-        char_print("\nYour current stats are:")
-        char_print(
-            f"Health: {self.health}/{self.max_health}"
-            f"\nAttack: {self.atk}"
-            f"\nDamage: {self.dmg}"
-            f"\nDefence: {self.dfc}"
-            f"\nPersuasion: {self.psd}"
-        )
-
-    def check_is_dead(self):
-        if self.health < 1:
-            self.run_death_sequence()
-            exit()
-
-    @staticmethod
-    def run_death_sequence():
-        char_print(
-            "\nClutching at your chest, your hands come away soaked in blood. Blinking, you stagger backwards and"
-            " keel over. Then, the world turns black."
-        )
-        char_input("\nYour story has ended. Enter anything to exit.")
-
-
-player_character = PlayerCharacter()
-
-
-@dataclass
-class PlayerInventory:
-    current_equipment = {}
-
-    def append_to_inventory(self, item: Item):
-        self.current_equipment.update({item.name.title(): item})
-
-    def remove_from_inventory(self, item: Item):
-        del self.current_equipment[item.name.title()]
-
-    def remove_random_item(self, player: PlayerCharacter = player_character):
-        if self.current_equipment:
-            to_discard = choice(list(self.current_equipment.values()))
-            if to_discard.is_equipped:
-                char_print("You've lost an equipped item!")
-                to_discard.is_equipped = False
-                if to_discard.type == "attire":
-                    player.dfc -= to_discard.dfc
-                    player.psd -= to_discard.psd
-                elif to_discard.type == "weapon":
-                    player.atk += to_discard.atk
-                    player.dmg += to_discard.dmg
-            del self.current_equipment[to_discard.name.title()]
-
-
-player_inventory = PlayerInventory()
